// features/custom-order/types.ts
export type OutdoorUse = 'moderee' | 'intensive'
export type DamasteelScale = 'large' | 'small'
// ...
export interface OutdoorForm {
  id: string
  label: string
  length: string
  description: string
  patternScale: DamasteelScale
  imageSrc?: string
  imageAlt?: string
  disabled?: boolean
}

export interface OutdoorUseCase {
  id: OutdoorUse
  label: string
  description: string
  note?: string
}